requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn
psycopg2-binary
sqlparse==0.4.2
whitenoise==6.2.0
django-filter
djangorestframework
flake8
tzdata
postgres
django_rest_passwordreset
PyYAML
flake8-bugbear
flake8-builtins
pep8-naming
flake8-commas
flake8-variables-names
requests
ujson
pytest-django
model_bakery
celery
redis
django-celery-beat
django-celery-results
python-decouple
faker
drf-spectacular
django-allauth
dj_database_url
selenium